Finally got the TAC to output aim info
NOTE: It took two tries over an entire week to notice, but this is the benefit of a second look (after sleep and doing other things). I, like the bitch I am, was trying to output a string as a %f and the UE4 compiler, like the bitch it is, kept outputting 0. I know I should expect 0,0,0 at the least. Anyway, small errors like this are a bitch.